OMv1 to legacy
