+++ scylla/tests/legacy/test_unit.py
@@ -10,28 +10,24 @@ from datadog_checks.base import AgentCheck
 from datadog_checks.base.errors import CheckException, ConfigurationError
 from datadog_checks.scylla import ScyllaCheck
 
-from .common import (
+from tests.common import (
     FLAKY_METRICS,
     INSTANCE_ADDITIONAL_GROUPS,
     INSTANCE_ADDITIONAL_METRICS,
-    INSTANCE_ADDITIONAL_METRICS_V2,
     INSTANCE_DEFAULT_GROUPS,
     INSTANCE_DEFAULT_METRICS,
-    INSTANCE_DEFAULT_METRICS_V2,
-    bucket_metrics,
     get_metrics,
-    transform_metrics_omv2,
 )
 
 
-@pytest.mark.skipif(PY2, reason='OpenMetrics V2 is only available with Python 3')
 @pytest.mark.unit
-def test_instance_default_check_omv2(aggregator, mock_db_data, dd_run_check, instance):
-    check = ScyllaCheck('scylla', {}, [instance])
+def test_instance_default_check(aggregator, mock_db_data, dd_run_check, instance_legacy):
+    check = ScyllaCheck('scylla', {}, [instance_legacy])
 
     dd_run_check(check)
     dd_run_check(check)
-    for m in INSTANCE_DEFAULT_METRICS_V2:
+
+    for m in INSTANCE_DEFAULT_METRICS:
         if m in FLAKY_METRICS:
             aggregator.assert_metric(m, count=0)
         else:
@@ -39,14 +35,12 @@ def test_instance_default_check_omv2(aggregator, mock_db_data, dd_run_check, ins
     aggregator.assert_all_metrics_covered()
 
 
-
-@pytest.mark.skipif(PY2, reason='OpenMetrics V2 is only available with Python 3')
 @pytest.mark.unit
-def test_instance_additional_check_omv2(aggregator, mock_db_data, dd_run_check, instance):
+def test_instance_additional_check(aggregator, mock_db_data, dd_run_check, instance_legacy):
     # add additional metric groups for validation
     additional_metric_groups = ['scylla.alien', 'scylla.sstables']
 
-    inst = deepcopy(instance)
+    inst = deepcopy(instance_legacy)
     inst['metric_groups'] = additional_metric_groups
 
     check = ScyllaCheck('scylla', {}, [inst])
@@ -55,30 +49,6 @@ def test_instance_additional_check_omv2(aggregator, mock_db_data, dd_run_check, 
     dd_run_check(check)
 
     metrics_to_check = get_metrics(INSTANCE_DEFAULT_GROUPS + additional_metric_groups)
-    transformed_metrics = transform_metrics_omv2(metrics_to_check) + bucket_metrics
-
-    for m in transformed_metrics:
-        if m in FLAKY_METRICS:
-            aggregator.assert_metric(m, count=0)
-        else:
-            aggregator.assert_metric(m)
-
-    aggregator.assert_all_metrics_covered()
-    aggregator.assert_service_check('scylla.openmetrics.health', status=AgentCheck.OK)
-
-
-@pytest.mark.skipif(PY2, reason='OpenMetrics V2 is only available with Python 3')
-@pytest.mark.unit
-def test_instance_full_additional_check_omv2(aggregator, mock_db_data, dd_run_check, instance):
-    inst = deepcopy(instance)
-    inst['metric_groups'] = INSTANCE_ADDITIONAL_GROUPS
-
-    check = ScyllaCheck('scylla', {}, [inst])
-
-    dd_run_check(check)
-    dd_run_check(check)
-
-    metrics_to_check = INSTANCE_DEFAULT_METRICS_V2 + INSTANCE_ADDITIONAL_METRICS_V2
 
     for m in metrics_to_check:
         if m in FLAKY_METRICS:
@@ -86,31 +56,50 @@ def test_instance_full_additional_check_omv2(aggregator, mock_db_data, dd_run_ch
         else:
             aggregator.assert_metric(m)
     aggregator.assert_all_metrics_covered()
-    aggregator.assert_service_check('scylla.openmetrics.health', status=AgentCheck.OK)
+    aggregator.assert_service_check('scylla.prometheus.health', status=AgentCheck.OK)
 
 
-@pytest.mark.skipif(PY2, reason='OpenMetrics V2 is only available with Python 3')
 @pytest.mark.unit
-def test_instance_invalid_group_check_omv2(aggregator, mock_db_data, dd_run_check, instance):
-    additional_metric_groups = ['scylla.bogus', 'scylla.sstables']
-
-    inst = deepcopy(instance)
-    inst['metric_groups'] = additional_metric_groups
+def test_instance_full_additional_check(aggregator, mock_db_data, dd_run_check, instance_legacy):
+    inst = deepcopy(instance_legacy)
+    inst['metric_groups'] = INSTANCE_ADDITIONAL_GROUPS
 
     check = ScyllaCheck('scylla', {}, [inst])
-    with pytest.raises(ConfigurationError):
-        check.load_configuration_models()
 
-    aggregator.assert_service_check('scylla.openmetrics.health', count=0)
+    dd_run_check(check)
+    dd_run_check(check)
+
+    metrics_to_check = INSTANCE_DEFAULT_METRICS + INSTANCE_ADDITIONAL_METRICS
+
+    for m in metrics_to_check:
+        if m in FLAKY_METRICS:
+            aggregator.assert_metric(m, count=0)
+        else:
+            aggregator.assert_metric(m)
+    aggregator.assert_all_metrics_covered()
+    aggregator.assert_service_check('scylla.prometheus.health', status=AgentCheck.OK)
 
 
-@pytest.mark.skipif(PY2, reason='OpenMetrics V2 is only available with Python 3')
 @pytest.mark.unit
-def test_invalid_instance_omv2(aggregator, mock_db_data, instance):
-    inst = deepcopy(instance)
-    inst.pop('openmetrics_endpoint')
+def test_instance_invalid_group_check(aggregator, mock_db_data, instance_legacy):
+    additional_metric_groups = ['scylla.bogus', 'scylla.sstables']
+
+    inst = deepcopy(instance_legacy)
+    inst['metric_groups'] = additional_metric_groups
+
+    with pytest.raises(ConfigurationError):
+        ScyllaCheck('scylla', {}, [inst])
+
+    aggregator.assert_service_check('scylla.prometheus.health', count=0)
+
+
+@pytest.mark.unit
+def test_invalid_instance(aggregator, mock_db_data, instance_legacy):
+    inst = deepcopy(instance_legacy)
+    inst.pop('prometheus_url')
 
     with pytest.raises(CheckException):
         ScyllaCheck('scylla', {}, [inst])
 
-    aggregator.assert_service_check('scylla.openmetrics.health', count=0)
+    aggregator.assert_service_check('scylla.prometheus.health', count=0)
+
